+++ next/src/pages/AboutUs.tsx
@@ -1,83 +1,64 @@
-import Sphere from '../components/aboutus/ReusableSpheres';
-import * as PageStyle from '../components/aboutus/AboutUs-Styled';
+import React from "react";
 
-const args1 = {
-    left: 11,
-    top: 7,
-    size: 13,
-    colourMain: "#969DC7",
-    colourSecondary: "#DAE9FB",
-    startMainPoint: -12,
-    startSecondaryPoint: 76.59,
-    angle: 261.11,
-    blur: 3.5,
-    rotation: 93.47,
-}
-
-const args2 = {
-    left: 45,
-    top: 23,
-    size: 9,
-    colourMain: "#D0E0ED",
-    colourSecondary: "#498AC1",
-    startMainPoint: 10.97,
-    startSecondaryPoint: 99.56,
-    angle: 261.11,
-    blur: 3,
-}
-
-const args3 = {
-    left: 14,
-    top: 62,
-    size: 9,
-    colourMain: "#9B9BE1",
-    colourSecondary: "#E8CAFF",
-    startMainPoint: -12,
-    startSecondaryPoint: 76.59,
-    angle: 261.11,
-    rotation: -74.2,
-}
-
-const args4 = {
-    left: 72,
-    top: 61,
-    size: 16,
-    colourMain: "#0069E7",
-    colourSecondary: "#BDDBFF",
-    startMainPoint: -10.14,
-    startSecondaryPoint: 81.0,
-    angle: 155.55,
-    rotation: 96.49,
-}
-
-const SphereArgs = [args1, args2, args3, args4];
-
-const CreateSpheres = SphereArgs.map((arg, index) => {
-    return (
-        <PageStyle.SpherePosition key={index} left={arg.left} top={arg.top}>
-            <Sphere {...arg} />
-        </PageStyle.SpherePosition>
-    )
-})
-
-const AboutUs = () => (
+export default function AboutUs() {
+  return (
     <div>
-        <PageStyle.AboutUsPage>
-            <PageStyle.AboutUsContent>
-                <PageStyle.AboutUsText>
-                    About Us
-                    <PageStyle.MainText>
-                        We are one of the biggest and most active societies at
-                        <PageStyle.BlueText> UNSW</PageStyle.BlueText>
-                        , catering to over
-                        <PageStyle.BlueText> 3500 CSE students </PageStyle.BlueText>
-                        spanning across degrees in Computer Science, Software Engineering, Bioinformatics and Computer Engineering.
-                    </PageStyle.MainText>
-                </PageStyle.AboutUsText>
-            </PageStyle.AboutUsContent>
-            {CreateSpheres}
-        </PageStyle.AboutUsPage>
-    </div>
-)
+      <h1>About</h1>
+      <p>
+        CSESoc is the official representative body of computing students at
+        UNSW. We are one of the largest and most active societies at UNSW, and
+        the largest computing society in the southern hemisphere. CSESoc
+        comprises ~9,500 UNSW students spanning across degrees in Computer
+        Science, Software Engineering, Bioinformatics and Computer Engineering.
+        We are here to fulfil the social, personal and professional needs of CSE
+        students, and promote computing through a variety of forms.
+      </p>
+      <p>
+        We are a society for the students, by the students. Here’s an overview
+        of what we do;
+      </p>
+      <ul>
+        <li>
+          Run weekly social and educational events, including trivia, movie,
+          boardgames nights, LAN parties, workshops, coding competitions, tech
+          talks, and our famous free weekly BBQ.
+        </li>
+        <li>
+          Create original media content, including Podcasts, articles, YouTube
+          videos, and live streams
+        </li>
+        <li>
+          Run a highly successful First Year Camp and Peer Mentoring program,
+          offering new CSE students (both undergraduate and postgraduate) a
+          chance to meet and mingle with other newcomers
+        </li>
+        <li>
+          Engage students with industry sponsors and representatives to develop
+          their professional capacity and curiosity
+        </li>
+        <li>
+          Develop our own open-source projects for students to get learn new
+          skills and develop tools for our community
+        </li>
+        <li>
+          Facilitate an online community of ~3k Discord users, ~5k Facebook
+          followers, ~600 YouTube subs, and ~500 Instagram followers
+        </li>
+      </ul>
 
-export default AboutUs
+      <h1>2022 Statistics</h1>
+      <ul style={{ listStyleType: "none", paddingLeft: "0" }}>
+        <li>🥳 100+ events (more on the way!) 🥳</li>
+        <li>📸 40+ media articles, podcast, videos, streams 📸</li>
+        <li>💸 32 sponsors 💸</li>
+        <li>💬 400 000 discord messages 💬</li>
+        <li>✨ 190 volunteers ✨</li>
+        <li>📼 40 000 Youtube views 📼</li>
+        <li>📼 600+ Youtube Subs 📼</li>
+        <li>🚸 500+ high school students reached 🚸</li>
+        <li>🧥374 hoodies 🧥</li>
+        <li>😷 250 face masks 😷</li>
+      </ul>
+    </div>
+  );
+}
