+++ next/src/pages/MiniEvents.tsx
