+++ next/src/pages/MiniHomepage.tsx
